% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_update.R
\name{adaptive_update}
\alias{adaptive_update}
\title{This function is a simple implementation of the adaptive MCMC algorithm of
Haario et al. (2001)}
\usage{
adaptive_update(
  chain,
  i,
  start_index = 1000,
  initial_Cd = 0.01,
  distribution = c("gamma", "gaussian")
)
}
\arguments{
\item{chain}{the current markov chain to update}

\item{i}{current chain iteration}

\item{start_index}{when should the proposal start adapting?}

\item{initial_Cd}{initial guess at the proposal variance}
}
\value{
a list with a whole bunch of stuff in it.
}
\description{
This function is a simple implementation of the adaptive MCMC algorithm of
Haario et al. (2001)
}
