% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.astroBayesPrediction.R
\name{plot.astroBayesPrediction}
\alias{plot.astroBayesPrediction}
\title{plots the age model with error bars for new predictions}
\usage{
\method{plot}{astroBayesPrediction}(age_predictions)
}
\arguments{
\item{age_predictions}{the output of \code{predict.astroBayesModel}

@import "tidyverse"}
}
\description{
plots the age model with error bars for new predictions
}
