% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_age_model.R
\name{plot_age_model}
\alias{plot_age_model}
\title{plots the age model}
\usage{
plot_age_model(age_model)
}
\arguments{
\item{age_model}{the output of \code{anchored_age_model}}
}
\value{
a list with a whole bunch of stuff in it.
}
\description{
plots the age model
}
