% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pgram.R
\name{plot_pgram}
\alias{plot_pgram}
\title{plots the best fit periodogram for each accumulation rate segment}
\usage{
plot_pgram(age_model)
}
\arguments{
\item{age_model}{the output of \code{anchored_age_model}}
}
\value{
a plot
}
\description{
plots the best fit periodogram for each accumulation rate segment
}
