% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sed_rate.R
\name{plot_sed_rate}
\alias{plot_sed_rate}
\title{plots the sedimentation rate density for each accumulation rate segment}
\usage{
plot_sed_rate(age_model)
}
\arguments{
\item{age_model}{the output of \code{anchored_age_model}}
}
\value{
a plot
}
\description{
plots the sedimentation rate density for each accumulation rate segment
}
