% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trace.R
\name{plot_trace}
\alias{plot_trace}
\title{plots the MCMC trace for each accumulation rate segment}
\usage{
plot_trace(age_model)
}
\arguments{
\item{age_model}{the output of \code{anchored_age_model}}
}
\value{
a plot
}
\description{
plots the MCMC trace for each accumulation rate segment
}
