% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.astroBayesModel.R
\name{predict.astroBayesModel}
\alias{predict.astroBayesModel}
\title{This function uses the output of an \code{\link{astro_bayes_model}} to predict the
age of new stratigraphic positions specified in \code{new_positions}}
\usage{
\method{predict}{astroBayesModel}(age_model, new_positions)
}
\arguments{
\item{age_model}{the output of \code{\link{astro_bayes_model}}}

\item{new_positions}{a data frame containing the name, position and positional
uncertainty (thickness) of stratigraphic points to predict the age of.
column headers \emph{must} be named exactly as follows:
\itemize{
\item id: name
\item position: stratigraphic position. The position must fall within the original evaluation
range of \code{age_model}
\item thickness: positional uncertainty is treated as a uniform distributions where the total
range equals position ± thickness ÷ 2
}}
}
\description{
This function uses the output of an \code{\link{astro_bayes_model}} to predict the
age of new stratigraphic positions specified in \code{new_positions}
}
