% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.astroBayesModel.R
\name{summary.astroBayesModel}
\alias{summary.astroBayesModel}
\title{This function uses the output of an \code{\link{astro_bayes_model}} to summarize
a variety of model parameters.
s}
\usage{
\method{summary}{astroBayesModel}(age_model, type)
}
\arguments{
\item{age_model}{the output of \code{\link{astro_bayes_model}}}

\item{type}{one of:
\itemize{
\item \code{sed_rate} calculate the posterior 95\% credible interval of sedimentation rate for each layer
\item \code{dates}    calculate the posterior 95\% credible interval of the dated horizons
\item \code{hiatus}   calculate the posterior 95\% credible interval of hiatus duration(s)
\item \code{anchor}   calculate the posterior 95\% credible interval of the anchor point age
}}
}
\description{
This function uses the output of an \code{\link{astro_bayes_model}} to summarize
a variety of model parameters.
s
}
