% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_opt_likelihood.R
\name{time_opt_likelihood}
\alias{time_opt_likelihood}
\title{Haario et al. (2001)}
\usage{
time_opt_likelihood(
  cyclostrat_data,
  sed_rate,
  target_frequency,
  f_low = 0.035,
  f_high = 0.065,
  roll = 10^3
)
}
\value{
a a proposed value from \code{chain}
}
\description{
Haario et al. (2001)
}
