% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_random_normal.R
\name{truncated_random_normal}
\alias{truncated_random_normal}
\title{Truncated Random Normal}
\usage{
truncated_random_normal(mean, sd, low = 1e-10, high = 1e+10)
}
\arguments{
\item{mean}{mean value}

\item{sd}{standard deviation}

\item{low}{lower truncation value}

\item{high}{upper truncation value}
}
\value{
a random number
}
\description{
Truncated Random Normal
}
