% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/truncated_standard_normal.R
\name{truncated_standard_normal}
\alias{truncated_standard_normal}
\title{Truncated Standard Normal}
\usage{
truncated_standard_normal(a, b)
}
\arguments{
\item{a}{lower truncation}

\item{b}{upper truncation}
}
\value{
a random number from a normal distribution with a mean of 0 and standard
deviation of 1.
}
\description{
Truncated Standard Normal
}
