% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_likelihood.R
\name{visualize_likelihood}
\alias{visualize_likelihood}
\title{helper function to visualize the likelihood functions given a model layering scheme}
\usage{
visualize_likelihood(
  layer_boundaries,
  cyclostrat_data,
  target_frequency,
  resolution = 0.1,
  method = "malinverno",
  plot = TRUE
)
}
\description{
helper function to visualize the likelihood functions given a model layering scheme
}
