% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptive_update.R
\name{adaptive_update}
\alias{adaptive_update}
\title{This function is a simple implementation of the adaptive MCMC algorithm of
Haario et al. (2001)}
\usage{
adaptive_update(
  chain,
  i,
  start_index = 1000,
  initial_Cd = 0.01,
  lower = -1e+10,
  upper = 1e+10
)
}
\arguments{
\item{chain}{the current markov chain to update}

\item{i}{current chain iteration}

\item{start_index}{when should the proposal start adapting?}

\item{initial_Cd}{initial guess at the proposal variance}

\item{lower}{lower truncation for the random normal distribution. Defaults to \code{-1E-10}}

\item{upper}{upper truncation for the random normal distribution. Defaults to \code{1E-10}}
}
\value{
a a proposed value from \code{chain}
}
\description{
This function is a simple implementation of the adaptive MCMC algorithm of
Haario et al. (2001)
}
