% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astro_bayes_model.R
\name{astro_bayes_model}
\alias{astro_bayes_model}
\title{Run a modified version of the algorithm from Malinverno et al., 2010 with an additional
modified to anchor the resulting age model using radioisotopic (e.g., U-Pb Ar/Ar) geochronology}
\usage{
astro_bayes_model(
  geochron_data,
  cyclostrat_data,
  tuning_frequency,
  segment_edges,
  sed_prior_range,
  iterations = 1e+05,
  burn = 5000
)
}
\arguments{
\item{geochron_data}{data frame containing radiometric age determinations, their uncertainties and stratigraphic positions, and a identifier variable.
column headers \emph{must} be named exactly as follows:
\itemize{
\item age: the weighted mean age in Ma
\item age_sd: 1 sigma uncertainty in Ma
\item position: stratigraphic position in depth coordinates (i.e. the top is 0)
\item id: sample name. currently only one sample per unique id is allowed.
}}

\item{cyclostrat_data}{a data frame where the first column is the stratigraphic position
and the second column is the cyclostratigraphic proxy record.
column headers \emph{must} be named exactly as follows:
\itemize{
\item position: stratigraphic position in depth coordinates (i.e. the top is 0).
\item value: the measurement value of the proxy record.
}}

\item{tuning_frequency}{data frame of tuning frequencies to use. Currently frequencies must be in cycles / Ma (i.e., long eccentricity ~ 1 / 0.405).
column headers \emph{must} be named exactly as follows:
\itemize{
\item frequency: the tuning frequency or frequencies to use
}}

\item{segment_edges}{stratigraphic points where sedimentation rate changes can  occur. Must be in the same stratigraphic scheme as geochron_data and cyclostrat_data.}

\item{sed_prior_range}{vector of 2 sedimentation rates that define the minimum
and maximum allowable sedimentation rates. These two rates define a uniform
prior distribution for sedimentation rate.}

\item{iterations}{how many Markov Chain Monte Carlo iterations should the model run for}

\item{burn}{how many initial iterations to toss when calculating credible intervals}
}
\value{
a list with a whole bunch of stuff in it.
}
\description{
Run a modified version of the algorithm from Malinverno et al., 2010 with an additional
modified to anchor the resulting age model using radioisotopic (e.g., U-Pb Ar/Ar) geochronology
}
