% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/astro_bayes_model.R
\name{astro_bayes_model}
\alias{astro_bayes_model}
\title{Run a modified version of the algorithm from Malinverno et al., 2010 with an additional
modified to anchor the resulting age model using radioisotopic (e.g., U-Pb Ar/Ar) geochronology}
\usage{
astro_bayes_model(
  geochron_data,
  cyclostrat_data,
  tuning_frequency,
  segment_edges,
  iterations = 10000,
  burn = 5000,
  method = NA
)
}
\arguments{
\item{geochron_data}{data frame containing radiometric age determinations, their uncertainties and stratigraphic positions, and a identifier variable.
column headers \emph{must} be named exactly as follows:
\itemize{
\item age: the weighted mean age in Ma
\item age_sd: 1 sigma uncertainty in Ma
\item position: stratigraphic position in depth coordinates (i.e. the top is 0)
\item id: sample name. currently only one sample per unique id is allowed.
}}

\item{cyclostrat_data}{a data frame where the first column is the stratigraphic position
and the second column is the cyclostratigraphic proxy record.
column headers \emph{must} be named exactly as follows:
\itemize{
\item position: stratigraphic position in depth coordinates (i.e. the top is 0).
\item value: the measurement value of the proxy record.
}}

\item{tuning_frequency}{data frame of tuning frequencies to use. Currently frequencies must be in cycles / Ma (i.e., long eccentricity ~ 1 / 0.405).
column headers \emph{must} be named exactly as follows:
\itemize{
\item frequency: the tuning frequency or frequencies to use
\item orbital_cycle: character string name of each orbital cycle.
Multiple rows with the same name are allowed.
}}

\item{segment_edges}{stratigraphic points where sedimentation rate changes can  occur. Must be in the same stratigraphic scheme as geochron_data and cyclostrat_data.}

\item{iterations}{how many Markov Chain Monte Carlo iterations should the model run for}

\item{burn}{how many initial iterations to toss when calculating credible intervals}
}
\value{
The function returns a list of class \code{astroBayesModel}
which contains the following objects:
\code{CI} data frame containing the credible interval for the age model
\code{anchor_point} a data.frame containing the posterior sample for the anchor_point
parameter(s)
\itemize{
\item \code{sed_rate} a matrix containing the posterior sample of sedimentation rate
for each model segment
\item \code{model_iterations} a matrix containing the individual model iterations used
to calculate \code{CI}. These can be plotted against the \code{position} column in \code{CI}
to visualize.
}

The output object also includes all the model inputs: shown below. These are
carried through for ease of data visualization.
\itemize{
\item \code{segment_edges}
\item \code{geochron_data}
\item \code{cyclostrat_data}
\item \code{sed_prior_range}
\item \code{tuning_frequency}
}
}
\description{
Run a modified version of the algorithm from Malinverno et al., 2010 with an additional
modified to anchor the resulting age model using radioisotopic (e.g., U-Pb Ar/Ar) geochronology
}
