% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.astroBayesModel.R
\name{plot.astroBayesModel}
\alias{plot.astroBayesModel}
\title{plots the age model}
\usage{
\method{plot}{astroBayesModel}(age_model)
}
\arguments{
\item{age_model}{the output of \code{astro_bayes_model}}
}
\value{
a list with a whole bunch of stuff in it.
}
\description{
plots the age model
}
