% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualize_likelihood.R
\name{visualize_likelihood}
\alias{visualize_likelihood}
\title{helper function to visualize the likeihood functions given a model segementaiton scheme}
\usage{
visualize_likelihood(
  sed_prior_range = c(0.1, 100),
  segment_edges,
  cyclostrat_data,
  tuning_frequency,
  resolution = 0.1,
  method = "malinverno"
)
}
\description{
helper function to visualize the likeihood functions given a model segementaiton scheme
}
